SeleniumWebDriverHelper and WebDriverWaitFactory moved to che.selenium.core
